ref: Move getWeeklySchedule read to server-only queries module
Server Actions ('use server') should only handle mutations. Moved the
read function and its types to queries.ts with import 'server-only',
keeping actions.ts for setDaySchedule and clearSchedule only.
