+++ app/(app)/flames/schedule/components/DayRow.tsx
@@ -16,7 +16,8 @@ import { toast } from 'sonner';
 import { Button } from '@/components/ui/button';
 import { cn, parseLocalDate } from '@/lib/utils';
 import type { Flame } from '@/utils/supabase/rows';
-import { type DayPlan, setDaySchedule } from '../actions';
+import { setDaySchedule } from '../actions';
+import type { DayPlan } from '../queries';
 import { ASSIGNED_FLAME_ZONE_ID, MY_FLAMES_ZONE_ID } from './constants';
 import { DayRowFuelBar } from './DayRowFuelBar';
 import { AssignedFlamesZone } from './dialog/AssignedFlamesZone';
